Implementado novas rotas e corrigido o bug no AutoMapper

+++ CadUserAPI/Controllers/UsuarioController.cs
@@ -7,7 +7,9 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.IdentityModel.Tokens;
+using Microsoft.Net.Http.Headers;
 using System;
+using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
@@ -36,19 +38,64 @@ namespace CadUserAPI.Application.Controllers
             _mapper = mapper;
         }
 
+        [HttpGet]
+        public async Task<ActionResult<IEnumerable<Usuario>>> Get()
+        {
+            var usuarios = await UsuarioRepository.GetAllAsync();
+            
+            if (!ValidateToken())
+            {
+                MensagemDTO mensagem = new MensagemDTO("Não autorizado!");
+                return Unauthorized(mensagem);
+            }
+
+            if (usuarios == null)
+            {
+                MensagemDTO mensagemDTO = new MensagemDTO("Operação inválida! Tente adicionar mais usuários!");
+                return BadRequest(mensagemDTO);
+            }
+
+            var usuarioResumoDTOs = _mapper.Map<IEnumerable<UsuarioResumoDTO>>(usuarios);
+            return Ok(usuarioResumoDTOs);
+        }
+
+
+        [HttpGet("{id}", Name = "ObterUsuario")]
+        public async Task<ActionResult<UsuarioDTO>> GetAsync(Guid id)
+        {
+            var usuario = await UsuarioRepository.GetAsync(id);
+
+            if (!ValidateToken())
+            {
+                MensagemDTO mensagem = new MensagemDTO("Não autorizado!");
+                return Unauthorized(mensagem);
+            }
+
+            if (usuario == null)
+            {
+                return NotFound();
+            }
+
+            var usuarioDTO = _mapper.Map<UsuarioDTO>(usuario);
+            return Ok(usuarioDTO);
+        }
+
         [AllowAnonymous]
         [HttpPost("login")]
         public async Task<IActionResult> Login([FromBody] LoginUserDTO userInfo)
         {
-            var user = await UsuarioRepository.GetByEmailAsync(userInfo.Email);
+            var user = await _userManager.FindByEmailAsync(userInfo.Email);
             if (user.Email != null)
             {
-                var result = await _signInManager.PasswordSignInAsync(userInfo.Email,
+                var result = await _signInManager.PasswordSignInAsync(user.Email,
                 userInfo.Password, isPersistent: false, lockoutOnFailure: false);
 
                 if (result.Succeeded)
                 {
-                    return Ok(GeraToken(user));
+                    var userDto = _mapper.Map<UsuarioDTO>(user);
+                    UsuarioTokenDTO aux = GeraToken(userDto);
+                    await UsuarioRepository.UpdateToken(user, aux.Token);
+                    return Ok(aux);
                 }
                 else
                 {
@@ -59,7 +106,7 @@ namespace CadUserAPI.Application.Controllers
             else
             {
                 MensagemDTO message = new MensagemDTO("Usuário e/ou senha inválidos");
-                return BadRequest(message);
+                return Unauthorized(message);
             }
 
         }
@@ -68,21 +115,22 @@ namespace CadUserAPI.Application.Controllers
         [HttpPost("registrar")]
         public async Task<IActionResult> Register([FromBody] RegisterUserDTO userInfo)
         {
+            if (await UsuarioRepository.GetByEmailAsync(userInfo.Email) != null)
+            {
+                MensagemDTO mensagem = new MensagemDTO("Já existe um usuário com esse email cadastrado!");
+                return Conflict(mensagem);
+            }
+
             var user = new Usuario
             {
                 UserName = userInfo.Email,
                 Name = userInfo.Name,
                 Email = userInfo.Email,
                 EmailConfirmed = true,
+                DDD = userInfo.Phone.DDD,
                 PhoneNumber = userInfo.Phone.Number,
                 PhoneNumberConfirmed = true
             };
-
-            if (await UsuarioRepository.GetByEmailAsync(user.Email) != null)
-            {
-                MensagemDTO mensagem = new MensagemDTO("Já existe um usuário com esse email cadastrado!");
-                return Conflict(mensagem);
-            }
 
             var result = await _userManager.CreateAsync(user, userInfo.Password);
 
@@ -94,7 +142,10 @@ namespace CadUserAPI.Application.Controllers
 
             await _signInManager.SignInAsync(user, false);
 
-            var aux = GeraToken(user);
+            var userDto = _mapper.Map<UsuarioDTO>(user);
+
+            UsuarioTokenDTO aux = GeraToken(userDto);
+            await UsuarioRepository.UpdateToken(user, aux.Token);
 
             return Ok(aux);
         }
@@ -114,6 +165,12 @@ namespace CadUserAPI.Application.Controllers
             {
                 MensagemDTO message = new MensagemDTO("Erro na alteração: ids diferentes!");
                 return BadRequest(message);
+            }
+
+            if (!ValidateToken())
+            {
+                MensagemDTO mensagem = new MensagemDTO("Não autorizado!");
+                return Unauthorized(mensagem);
             }
 
             try
@@ -137,6 +194,12 @@ namespace CadUserAPI.Application.Controllers
                 return NotFound();
             }
 
+            if (!ValidateToken())
+            {
+                MensagemDTO mensagem = new MensagemDTO("Não autorizado!");
+                return Unauthorized(mensagem);
+            }
+
             try
             {
                 await UsuarioRepository.DeleteAsync(id);
@@ -148,12 +211,40 @@ namespace CadUserAPI.Application.Controllers
             }
         }
 
-        private object GeraToken(Usuario userInfo)
+        public bool ValidateToken()
+        {
+            var accessToken = Request.Headers[HeaderNames.Authorization].ToString();
+            var token = accessToken.Substring(7, accessToken.Length - 7);
+
+            var tokenHandler = new JwtSecurityTokenHandler();
+            try
+            {
+                tokenHandler.ValidateToken(token, new TokenValidationParameters
+                {
+                    ValidateIssuer = true,
+                    ValidateAudience = true,
+                    ValidateLifetime = true,
+                    ValidAudience = _configuration["TokenConfiguration:Audience"],
+                    ValidIssuer = _configuration["TokenConfiguration:Issuer"],
+                    ValidateIssuerSigningKey = true,
+                    IssuerSigningKey = new SymmetricSecurityKey(
+                         Encoding.UTF8.GetBytes(_configuration["Jwt:key"]))
+                }, out SecurityToken validatedToken);
+            }
+            catch
+            {
+                return false;
+            }
+            return true;
+        }
+
+
+        private UsuarioTokenDTO GeraToken(UsuarioDTO userInfo)
         {
             var claims = new[]
             {
-                new Claim(JwtRegisteredClaimNames.Jti, userInfo.Id.ToString()),
-                new Claim(JwtRegisteredClaimNames.UniqueName, userInfo.Email)
+                new Claim(JwtRegisteredClaimNames.UniqueName, userInfo.Id.ToString()),
+                new Claim(JwtRegisteredClaimNames.Email, userInfo.Email)
             };
 
             var key = new SymmetricSecurityKey(
@@ -170,11 +261,9 @@ namespace CadUserAPI.Application.Controllers
                 expires: expiration,
                 signingCredentials: credentials);
 
-            var userDto = _mapper.Map<UsuarioDTO>(userInfo);
-
             return new UsuarioTokenDTO()
             {
-                User = userDto,
+                User = userInfo,
                 Authenticated = true,
                 Token = new JwtSecurityTokenHandler().WriteToken(token),
                 Expiration = expiration,
